feat : automatic selection of subtask
Adds the automatic selection of a subtask when a task is forced to have
one, and branches for UI whether or not the subtask can stay null.
Also localise the filter function in SubtaskSelect since it is only used
there, and add a if for the component in form to only load after a task
is selected.