Fix stat display (#19)
This pull request fixes the problem when displaying TaskRatioList because of a div that were too narrow, and fixes the fact that week bar chart for the year is narrower than the other charts. It also adds labels to the charts instead of using ApexCharts title, for commodity and readability reasons.