feat : improving Task and Tag modals
This commit add a new error code management in the Task and Tag modals,
taking into account a 422 return from the API in case of an already
existing ressource.
It also adds a validation for spaces and displays a new Info component
in case one is found, and some tweks on the spacing of UI elements.