refactor : generic stats and modifs
Created a generic methods for every period, passing the period and
optionally a date to reflect the changes in the API where it is now
centralised.
Updated the parsing of the response to match the new response needed in
case a past stat is asked : now 'resume' is the equivalent of the home
stats stored in the store for all current periods since it needs to be
given in the response, and all former fields are in the 'details'
object.