feat : adds date parameter to GenericStats
This commit make possible to request generic stats for past classical
time periods (day, week, month, year). The Views are modified
accordingly, and QueryForm now sends change events each time its value
changes so that the GenericStat elements get unmounted from the views
and mounted back with its new API calls.
It also does some cleaning to the statStore : individual generic stat
functions are discarded in favor of a single one with parameters, and
one is added to get the past task time ratios.
The coupling with the statStore in the DOM is also modified.