fix : ensuring good updates of from values at creationg from modal
There were issues when creating a task or a tag : either they were displayed but not updated as data, either updated as data but not displayed.
Also the tag stored to be displayed contained the guid, which we didn't want and was related to the way createdTag was updated in the tag store.