refactor : menus
This commit allows the StickyMenu to be fixed on the screen and still allow the overlay to cover every view by extracting it into the App.vue file.
It take some props and emits to make the components and the view communicate, which have been added, and I splitted some logic for clarity.