style : apex charts color
Copy pasted the three first colors from the apexCharts library doc so that the pages get homogeneous colors.
Might create a little friction if the colors of the tasks in CustomBar/TaskRatioList differ from the color the same tasks have in the chat, but that a minor inconvenience I'll take.
Also fixed the unreadable menus and tooltips from apexcharts components.